+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Milliliters is now supported for custom brew batches
+- Minimal Button Toggle for v1.0.4 button style
 
 ## [1.0.4] - 2024-09-10
 ### Changes